feat: Optimize image loading, fix middleware issue, and update API base URL and Cypress configuration
