+++ app/business/user/user.query.ts
@@ -4,19 +4,6 @@ import axios from 'axios';
 import { getToken } from '../auth';
 import { isValidation } from '@/app/utils/zod/validation.util';
 import { UserInfoResponseSchema, InitUserInfoResponseSchema } from './user.validation';
-import { UnauthorizedError } from '@/app/utils/http/http-error';
-
-export async function auth(): Promise<InitUserInfoResponse | UserInfoResponse | undefined> {
-  try {
-    const result = await fetchUserInfo();
-    return result;
-  } catch (error) {
-    if (error instanceof UnauthorizedError) {
-      return;
-    }
-    throw error;
-  }
-}
 
 export async function fetchUserInfo() {
   try {
