+++ middleware.ts
@@ -1,6 +1,6 @@
 import type { NextRequest } from 'next/server';
 import { validateToken } from './app/business/user/user.command';
-import { fetchUserInfo } from './app/business/user/user.query';
+import { fetchUser } from './app/business/auth';
 
 async function getAuth(request: NextRequest): Promise<{
   role: 'guest' | 'user' | 'init';
@@ -24,7 +24,7 @@ async function getAuth(request: NextRequest): Promise<{
 
   request.cookies.set('accessToken', validatedResult.accessToken);
 
-  const user = await fetchUserInfo();
+  const user = await fetchUser();
   return {
     role: user.studentName ? 'user' : 'init',
   };
@@ -43,21 +43,18 @@ function isAllowedGuestPath(path: string, strict: boolean = false) {
 }
 
 export async function middleware(request: NextRequest) {
-  const isAuth = request.nextUrl.searchParams.get('isAuth');
+  const auth = await getAuth(request);
 
-  if (isAuth === 'true') {
-    const auth = await getAuth(request);
-    if (auth.role === 'init' && !request.nextUrl.pathname.startsWith('/grade-upload')) {
-      return Response.redirect(new URL('/grade-upload', request.url));
-    }
+  if (auth.role === 'init' && !request.nextUrl.pathname.startsWith('/grade-upload')) {
+    return Response.redirect(new URL('/grade-upload', request.url));
+  }
 
-    if (auth.role === 'guest' && !isAllowedGuestPath(request.nextUrl.pathname)) {
-      return Response.redirect(new URL('/sign-in', request.url));
-    }
+  if (auth.role === 'guest' && !isAllowedGuestPath(request.nextUrl.pathname)) {
+    return Response.redirect(new URL('/sign-in', request.url));
+  }
 
-    if (auth.role !== 'guest' && isAllowedGuestPath(request.nextUrl.pathname, true)) {
-      return Response.redirect(new URL('/my', request.url));
-    }
+  if (auth.role !== 'guest' && isAllowedGuestPath(request.nextUrl.pathname, true)) {
+    return Response.redirect(new URL('/my', request.url));
   }
 }
